Add two new control parameters.
Curve: Index into a selection of some (four) preconfigured blends of polynomial curves. More can be added!
Threshold: A factor for the above blending process (dynamic so it reacts to changes during processing). Does not apply to all blends though. This is noted in the label in the generic UI.